Update LeetCode #142 Linked List Cycle II.py
Solution
Use two pointer with different speed, say runner and walker, and make runner be twice faster than walker. 
If there is no cycle in this linked list, runner will finally run through entire list without meet walker.
If there exists a cycle, they will meet each other at a node, say x, and x represents the x-th node in this list. 
Let’s assume the position of cycle start is at y-th node and the length of cycle is m.
About runner, it has already run 2x nodes, which is equal to y + m + (x — y) (length to cycle start + length of cycle + nodes to meet walker). 
So we have 2x = y + m + (x — y), and it can conduct to x = m.
About walker, it needs to walk more m — (x — y) = m — x + y nodes to reach cycle start. As we know in 1 that x is equal to m, 
walker needs to walk more y nodes to reach the node of cycle start.
By 1 and 2, we can put another pointer, say seeker, with the same speed as walker at the head.
Because seeker also needs y move nodes to reach the node of cycle start, the node that seeker and walker meet each other will be what we are looking for.